+++ R/coef_extractor.R
@@ -25,12 +25,14 @@ coefficient_extractor <- function(object,...){
 
 #' @rdname coefficient_extractor
 #' @param component Which component should be extracted
+#' @param exclude_degenerate Drop bootstrap replicates flagged degenerate
+#'   (default \code{TRUE}); see \code{\link{evinf_control}}.
 #' @export
-coefficient_extractor.evzinb <- function(object, component = c('all','count','zero','evi','pareto'), ...){
+coefficient_extractor.evzinb <- function(object, component = c('all','count','zero','evi','pareto'), exclude_degenerate = TRUE, ...){
 
   component <- normalize_component(component, c('all','count','zero','evi','pareto'))
 
-  object$bootstraps <- object$bootstraps %>% purrr::discard(~'try-error' %in% class(.x))
+  object$bootstraps <- evinf_usable_bootstraps(object, exclude_degenerate)
 
   count_boot <- object$bootstraps %>% purrr::map('coef') %>% purrr::map('Beta.NB') %>% dplyr::bind_rows()
   zero_boot <- object$bootstraps %>% purrr::map('coef') %>% purrr::map('Beta.multinom.ZC') %>% dplyr::bind_rows()
@@ -55,11 +57,11 @@ coefficient_extractor.evzinb <- function(object, component = c('all','count','ze
 
 #' @rdname coefficient_extractor
 #' @export
-coefficient_extractor.evinb <- function(object, component = c('all','count','evi','pareto'), ...){
+coefficient_extractor.evinb <- function(object, component = c('all','count','evi','pareto'), exclude_degenerate = TRUE, ...){
 
   component <- normalize_component(component, c('all','count','evi','pareto'))
 
-  object$bootstraps <- object$bootstraps %>% purrr::discard(~'try-error' %in% class(.x))
+  object$bootstraps <- evinf_usable_bootstraps(object, exclude_degenerate)
 
   count_boot <- object$bootstraps %>% purrr::map('coef') %>% purrr::map('Beta.NB') %>% dplyr::bind_rows()
   evi_boot <- object$bootstraps %>% purrr::map('coef') %>% purrr::map('Beta.multinom.PL') %>% dplyr::bind_rows()